requirements.txt
SQLAlchemy==1.4.25
Flask_SQLAlchemy==2.5.1
PyMySQL==1.0.2
Werkzeug==1.0.1
Flask==1.1.2
cryptography==35.0.0
protobuf==3.18.1
python-dotenv==0.19.1
Random-Word-Generator==1.3
google-cloud-storage==1.42.3
